+++ README.md
@@ -18,16 +18,6 @@
 
 A comprehensive monorepo for PostgreSQL Abstract Syntax Tree (AST) parsing, manipulation, and code generation. This collection of packages provides everything you need to work with PostgreSQL at the AST level, from parsing SQL queries to generating type-safe TypeScript definitions.
 
-## 📦 Packages 
-
-| Package | Description | Key Features |
-|---------|-------------|--------------|
-| [**pgsql-parser**](./packages/parser) | The real PostgreSQL parser for Node.js | • Uses actual PostgreSQL C parser via WebAssembly<br>• Symmetric parsing and deparsing<br>• Battle-tested with 23,000+ SQL statements |
-| [**pgsql-deparser**](./packages/deparser) | Lightning-fast SQL generation from AST | • Pure TypeScript, zero runtime dependencies<br>• No WebAssembly overhead<br>• Perfect for AST-to-SQL conversion only |
-| [**@pgsql/cli**](./packages/pgsql-cli) | Unified CLI for all PostgreSQL AST operations | • Parse SQL to AST<br>• Deparse AST to SQL<br>• Generate TypeScript from protobuf<br>• Single tool for all operations |
-| [**@pgsql/utils**](./packages/utils) | Type-safe AST node creation utilities | • Programmatic AST construction<br>• Runtime Schema<br>• Seamless integration with types |
-| [**pg-proto-parser**](./packages/proto-parser) | PostgreSQL protobuf parser and code generator | • Generate TypeScript interfaces from protobuf<br>• Create enum mappings and utilities<br>• AST helper generation |
-
 ## 🚀 Quick Start
 
 ### Installation
@@ -70,6 +60,41 @@ const sql = await deparse(ast);
 console.log(sql); // SELECT * FROM users WHERE id = 1
 ```
 
+#### Build AST with Types
+```typescript
+import { deparse } from 'pgsql-deparser';
+import { SelectStmt } from '@pgsql/types';
+
+const stmt: { SelectStmt: SelectStmt } = {
+    SelectStmt: {
+        targetList: [
+            {
+                ResTarget: {
+                    val: {
+                        ColumnRef: {
+                            fields: [{ A_Star: {} }]
+                        }
+                    }
+                }
+            }
+        ],
+        fromClause: [
+            {
+                RangeVar: {
+                    relname: 'some_table',
+                    inh: true,
+                    relpersistence: 'p'
+                }
+            }
+        ],
+        limitOption: 'LIMIT_OPTION_DEFAULT',
+        op: 'SETOP_NONE'
+    }
+};
+
+await deparse(stmt);
+```
+
 #### Build AST Programmatically
 ```typescript
 import * as t from '@pgsql/utils';
@@ -98,19 +123,16 @@ const stmt: { SelectStmt: SelectStmt } = t.nodes.selectStmt({
 await deparse(stmt);
 ```
 
-#### Use the CLI
-```bash
-npm install -g @pgsql/cli
+## 📦 Packages 
 
-# Parse SQL file
-pgsql parse query.sql
+| Package | Description | Key Features |
+|---------|-------------|--------------|
+| [**pgsql-parser**](./packages/parser) | The real PostgreSQL parser for Node.js | • Uses actual PostgreSQL C parser via WebAssembly<br>• Symmetric parsing and deparsing<br>• Battle-tested with 23,000+ SQL statements |
+| [**pgsql-deparser**](./packages/deparser) | Lightning-fast SQL generation from AST | • Pure TypeScript, zero runtime dependencies<br>• No WebAssembly overhead<br>• Perfect for AST-to-SQL conversion only |
+| [**@pgsql/cli**](./packages/pgsql-cli) | Unified CLI for all PostgreSQL AST operations | • Parse SQL to AST<br>• Deparse AST to SQL<br>• Generate TypeScript from protobuf<br>• Single tool for all operations |
+| [**@pgsql/utils**](./packages/utils) | Type-safe AST node creation utilities | • Programmatic AST construction<br>• Runtime Schema<br>• Seamless integration with types |
+| [**pg-proto-parser**](./packages/proto-parser) | PostgreSQL protobuf parser and code generator | • Generate TypeScript interfaces from protobuf<br>• Create enum mappings and utilities<br>• AST helper generation |
 
-# Convert AST to SQL
-pgsql deparse ast.json
-
-# Generate TypeScript from protobuf
-pgsql proto-gen --inFile pg_query.proto --outDir out --types --enums
-```
 
 ## 🛠️ Development
